rename e2e folder
