+++ test/new-e2e/tests/agent-configuration/sgc-embedded/fixtures/secrets.yaml
