+++ test/new-e2e/tests/agent-configuration/sgc-embedded/sgc_common_test.go
@@ -3,8 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-// package secretbackend contains e2e tests for secret management (runtime)
-package secretbackend
+// package sgcbackend contains e2e tests for secret management (runtime)
+package sgcbackend
 
 import (
 	"github.com/DataDog/datadog-agent/test/e2e-framework/testing/e2e"
